+++ backend/canisters/group/c2c_client/src/lib.rs
@@ -20,7 +20,6 @@ generate_c2c_call!(c2c_bot_delete_messages);
 generate_c2c_call!(c2c_bot_invite_users);
 generate_c2c_call!(c2c_bot_remove_user);
 generate_c2c_call!(c2c_bot_send_message);
-generate_c2c_call!(c2c_bot_subscribe_to_events);
 generate_c2c_call!(c2c_delete_group);
 generate_c2c_call!(c2c_export_group);
 generate_c2c_call!(c2c_export_group_events);