+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Send BotEventWrapper as HTTP body & signature as a header ([#8570](https://github.com/open-chat-labs/open-chat/pull/8570))
+- Remove endpoint for manual bot event subscriptions ([#8594](https://github.com/open-chat-labs/open-chat/pull/8594))
 
 ## [[2.0.1879](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1879-local_user_index)] - 2025-08-27